Decouple TcpEndpoint target type
